backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
asyncpg>=0.29
python-dotenv>=1.0
pydantic>=2.7
httpx>=0.27
pypdf>=4.2
python-multipart>=0.0.9
google-genai>=1.33.0

